apps/backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.7
numpy>=1.24.3  # Allow newer compatible versions
pandas>=2.0.3  # Allow newer compatible versions
python-dotenv==1.0.0
Flask-CORS==4.0.0
Werkzeug==2.3.7
Flask-Login==0.6.2
Flask-Session==0.5.0  # Redis-backed session management
# reportlab==4.0.7
Flask-Mail==0.9.1
Authlib==1.2.1
# Data processing and scraping
requests==2.31.0
selenium==4.15.2
beautifulsoup4==4.12.2

# Enterprise enhancements
cryptography==41.0.4
PyJWT==2.8.0  # For advanced session management

celery==5.3.4  # Background task processing

# Data validation and monitoring
marshmallow==3.20.1
structlog==23.2.0  # Structured logging
psutil==5.9.6  # System monitoring

# Development and testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
pytest-mock==3.12.0
flake8==6.1.0
black==23.11.0
isort==5.12.0
mypy==1.7.1
bandit==1.7.5
safety>=3.0.0
coverage==7.3.2
